requirements.txt
streamlit
langchain
faiss-cpu
pypdf
langchain-groq
sentence-transformers
python-dotenv